Optimize Update(Chunk).
Rewrite Update(Chunk) to avoid as much as possible the multiple calls of
Update(Byte).
This optimization corrects one of the negative effect of the previous commit on
the use case 3

Time used to process ~440MiB of data with SipHash-2-4
-----------+------------+------------
| Use case | Before (s) | After (s) |
-----------+------------+------------
|    1     |   5.156    |   5.090   |
|    2     |   1.148    |   1.134   |
|    3     |   4.791    |   0.487   |
-------------------------------------

As a reminder, the use case 3 is: hash the data incrementally (Streaming mode),
by chunk (of 4096 bytes for example).The chunks' sizes can be variable.